% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runHMC.R
\name{compute_loglik}
\alias{compute_loglik}
\title{Little function to compute the log-likelihood (for the obs vs censored cases)}
\usage{
compute_loglik(f, s)
}
\arguments{
\item{f}{The output from internal \code{survHEhmc} function that uses the model
output and computes the log-density for models that do not need to separate
out the observed and censored data}

\item{s}{The output from internal \code{survHEhmc} function that uses the model
output and computes the log-density for models that do need to separate
out the observed and censored data}
}
\description{
Little function to compute the log-likelihood (for the obs vs censored cases)
}
\references{
Baio (2020). survHE
}
\author{
Gianluca Baio
}
\keyword{Carlo}
\keyword{Hamiltonian}
\keyword{Monte}
\keyword{Parametric}
\keyword{models}
\keyword{survival}
