% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survHEhmc-package.R
\docType{package}
\name{survHEhmc-package}
\alias{survHEhmc}
\alias{survHEhmc-package}
\title{survHEhmc: Survival Analysis in Health Economic Evaluation using Hamiltonian Monte Carlo}
\description{
A module to complement the backbone structure of the package 'survHE' and expand its functionality to run survival models under a Bayesian approach (based on Hamiltonian Monte Carlo). \doi{10.18637/jss.v095.i14}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/giabaio/survHEhmc/}
  \item \url{https://gianluca.statistica.it/software/survhe/}
  \item Report bugs at \url{https://github.com/giabaio/survHEhmc/issues}
}

}
\author{
\strong{Maintainer}: Gianluca Baio \email{g.baio@ucl.ac.uk}

Other contributors:
\itemize{
  \item Philip Cooney \email{philip.cooney@hotmail.com} [contributor]
  \item Andrew Jones \email{andrew.r.johnson@postgrad.curtin.edu.au} [contributor]
}

}
\keyword{internal}
